refactor(key_base): 优化按键处理流程并修复状态同步问题
- 合并三次链表遍历为单次遍历，提高处理效率
- 修复 diff_timer 为0时的状态不同步问题
- 缓存当前时间值减少函数调用开销
- 调整消抖计数器类型为 uint16_t 以支持更大范围
- 优化代码结构，减少重复检查

+++ Q16_Drive_Opam_clion/modules/key_base/key_base.h
@@ -124,8 +124,8 @@ struct key_base_context {
     uint8_t combination_handled : 1;    /**< 标记组合事件已处理 */
     uint8_t combination_long_handled : 1; /**< 标记组合长按事件已处理 */
     uint8_t batter_counts;              /**< 按键点击计数 */
-    uint8_t press_debounce_count;       /**< 按键按下消抖计数 */
-    uint8_t release_debounce_count;     /**< 按键释放消抖计数 */
+    uint16_t press_debounce_count;       /**< 按键按下消抖计数 */
+    uint16_t release_debounce_count;     /**< 按键释放消抖计数 */
 
     uint8_t key_event : 4;              /**< 当前按键事件 */
     uint8_t last_key_event : 4;         /**< 上次按键事件 */
